QoS Region pinger was configured to work only on .net 4.5 and .net core 2.0, but was being built for netstandard 2.0. Update #if so that .net standard 1.1 is explcitly excluded and all other .net standard, .net framework, and .net core are included.
The classes are only supported in .net standard 2.0 and newer (or any full framework above 1.1 and any .net core), but I assume we will never support anything between .net standard 1.1 and 2.0. If we do it will fail to compile which will be an obvious sign to explicitly exclude it here.